fix(backup): widen statvfs fields to u64 for macOS
statvfs f_bavail/f_blocks/f_frsize are u32 on macOS but u64 on Linux; multiply after widening so the backup-space check compiles on Apple targets. Unblocks the macOS release lane.

+++ crates/focusa-core/src/runtime/backup_io.rs
@@ -317,8 +317,8 @@ pub(super) fn filesystem_space(path: &Path) -> Result<(u64, u64)> {
         return Err(std::io::Error::last_os_error().into());
     }
     Ok((
-        stat.f_bavail.saturating_mul(stat.f_frsize),
-        stat.f_blocks.saturating_mul(stat.f_frsize),
+        u64::from(stat.f_bavail).saturating_mul(u64::from(stat.f_frsize)),
+        u64::from(stat.f_blocks).saturating_mul(u64::from(stat.f_frsize)),
     ))
 }
 #[cfg(windows)]
